styles: Disable transitions on load
When switching to the user-preferred scheme in JavaScript on page load
it's not very pretty to see the flicker the transitions make. This
change ensures transitions are only enabled when <body> has a certain
class, and the JavaScript removes this class during the initial scheme
switch on page load.

There seems to also be an issue with async/deferred <script>'s. The
transitions are still happening when "defer" was added to the script
tag. This is unfortunate, as it causes more blocking while fetching
resources. The flicker the transitions make are worse though, so need to
live with this for now until a better solution presents itself.