styles: Better onload transition handling
Transitions were still broken with the previous change (and I expect
there to still be issues).

One issue was that the `body.transitions *` rule do not select `<body>`
itself, and so the main background wasn't properly animated. This then
defeats the purpose of having transitions in the first place, but did at
least avoid the initial flicker.

Adding `body.transitions` to the CSS selector for transitions caused
the background transition to yet again be visible on load. This happens
if transitions are enabled through the `transitions` class on `<body>`
before the `transition-delay` is over.

This change tries to calculate for how long transitions need to be
disabled while changing styles on page load, re-enabling transitions
once the transition is supposed to have been completed. Seems to work
in the first few manual tests of this...