ssg: Don't rewrite external URLs
URLs with a `.org` extension are rewritten to `.html` to make sure Org
links between articles point to the live posts. This obviously doesn't
work very well when linking to external `.org` top-level domains. This
change adds a check that ensures that only local links are rewritten,
not external ones (crude check for protocol separator "://").